[*] fix code syntax and rename file + fix version in package.json
