+++ package.json
@@ -23,7 +23,7 @@
     "bcryptjs": "2.4.3",
     "bluebird": "3.3.4",
     "body-parser": "1.15.0",
-    "esprima": "^3.1.3",
+    "esprima": "3.1.3",
     "express": "4.14.0",
     "express-cors": "0.0.3",
     "express-jwt": "3.3.0",
